added in some vault records

+++ src/main/java/org/devtop/resources/EncryptionResource.java
@@ -7,36 +7,21 @@ import jakarta.ws.rs.DELETE;
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.POST;
 import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
 import jakarta.ws.rs.Produces;
-import jakarta.ws.rs.QueryParam;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 import java.util.List;
 import org.devtop.encryption.AES256;
 import org.devtop.entity.KvEntity;
-import org.devtop.json.DeleteKv;
 import org.devtop.json.KeyValue;
 import org.eclipse.microprofile.config.inject.ConfigProperty;
-import org.jboss.logging.annotations.Param;
 
 @Path("")
 public class EncryptionResource {
 
     @ConfigProperty(name = "AES_SECRET")
     private String secret;
-
-    @GET
-    @Produces(MediaType.TEXT_PLAIN)
-    public String hello() {
-        return "Hello from Quarkus REST";
-    }
-
-    @GET
-    @Path("/test")
-    @Produces(MediaType.TEXT_PLAIN)
-    public String test() {
-        return "This is a test";
-    }
 
     @GET
     @Path("/secret")
@@ -65,7 +50,7 @@ public class EncryptionResource {
     @GET
     @Path("/kv")
     public List<KvEntity> getKv() {
-        return KvEntity.listAll();
+        return KvEntity.findDecrypted(secret);
     }
 
     @POST
@@ -85,17 +70,25 @@ public class EncryptionResource {
         }
 
     }
-    
+
     @DELETE
-    @Path("/kv/delete")
+    @Path("/kv/{id}")
     @Transactional
-    public Response deleteKv(DeleteKv kv){
-         try {
-            KvEntity kve     = new KvEntity();
-            KvEntity kvFound = kve.findById(kv.id);
-            kvFound.delete();
-            
-            return Response.ok().build();
+    public Response deleteKv(@PathParam("id") int id) {
+        try {
+            KvEntity kvFound = KvEntity.findById(id);
+
+            if (kvFound != null) {
+                kvFound.delete();
+                return Response.ok()
+                        .entity("Successfully deleted record")
+                        .build();
+            }
+
+            return Response.status(Response.Status.NOT_FOUND)
+                    .entity("Record not found")
+                    .build();
+
         } catch (Exception e) {
             return Response.serverError()
                     .entity("Failed to process request: " + e.getMessage())
